@@ -4,13 +4,22 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
+
+  http: ^1.2.2
+
+  flutter_spinkit: ^5.2.1
+
+  geolocator: ^13.0.1
+
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
+
+
+  cupertino_icons: ^1.0.8
 
 dev_dependencies:
   flutter_test:
