@@ -1,4 +1,10 @@
+import 'package:clima/screens/location_screen.dart';
 import 'package:flutter/material.dart';
+import 'package:clima/services/location.dart';
+import 'package:clima/services/networking.dart';
+import 'package:flutter_spinkit/flutter_spinkit.dart';
+
+const weather_key = "45db4ebed876abbd1ec6a6fc6397caec";
 
 class LoadingScreen extends StatefulWidget {
   @override
@@ -6,17 +12,44 @@ class LoadingScreen extends StatefulWidget {
 }
 
 class _LoadingScreenState extends State<LoadingScreen> {
+  double? latitude;
+  double? longitude;
+
+  void getLocationData() async {
+    Location location = Location();
+    await location.getCurrentLocation();
+
+    // Use the location instance directly to get latitude and longitude
+    latitude = location.latitude;
+    longitude = location.longitude;
+
+    NetworkHelper networkHelper =
+        NetworkHelper(latitude, longitude, weather_key);
+
+    var weatherData = await networkHelper.getData();
+
+    Navigator.push(context, MaterialPageRoute(builder: (context) {
+      return LocationScreen(locationWeather: weatherData);
+    }));
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    getLocationData();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: Center(
-        child: RaisedButton(
-          onPressed: () {
-            //Get the current location
-          },
-          child: Text('Get Location'),
+        child: SpinKitDoubleBounce(
+          color: Colors.white,
+          size: 100.0,
         ),
       ),
     );
   }
 }
+
+//
