@@ -1,12 +1,46 @@
+import 'dart:js_interop';
+
+import 'package:clima/services/weather.dart';
 import 'package:flutter/material.dart';
 import 'package:clima/utilities/constants.dart';
+import 'package:clima/screens/city_screen.dart';
 
 class LocationScreen extends StatefulWidget {
+  LocationScreen({this.locationWeather});
+
+  final locationWeather;
+
   @override
   _LocationScreenState createState() => _LocationScreenState();
 }
 
 class _LocationScreenState extends State<LocationScreen> {
+  WeatherModel weatherModel = WeatherModel();
+  double? temperature;
+  String? cityName;
+  late String message;
+  late int condition;
+  var weathericon;
+
+  void initState() {
+    super.initState();
+    updateUI(widget.locationWeather);
+  }
+
+  void updateUI(dynamic weatherData) {
+    var temp = weatherData['main']['temp'];
+    temperature = temp.toInt();
+
+    condition = weatherData['weather'][0]['id'];
+    weathericon = weatherModel.getWeatherIcon(condition);
+
+    cityName = weatherData['name'];
+
+    message = weatherModel.getMessage(temperature as int);
+
+    print("oga oga");
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -28,15 +62,20 @@ class _LocationScreenState extends State<LocationScreen> {
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: <Widget>[
-                  FlatButton(
+                  FloatingActionButton(
                     onPressed: () {},
                     child: Icon(
                       Icons.near_me,
                       size: 50.0,
                     ),
                   ),
-                  FlatButton(
-                    onPressed: () {},
+                  FloatingActionButton(
+                    onPressed: () async {
+                      var typedName = await Navigator.push(context,
+                          MaterialPageRoute(builder: (context) {
+                        return CityScreen();
+                      }));
+                    },
                     child: Icon(
                       Icons.location_city,
                       size: 50.0,
@@ -49,11 +88,11 @@ class _LocationScreenState extends State<LocationScreen> {
                 child: Row(
                   children: <Widget>[
                     Text(
-                      '32°',
+                      '$temperature°',
                       style: kTempTextStyle,
                     ),
                     Text(
-                      '☀️',
+                      '$weathericon',
                       style: kConditionTextStyle,
                     ),
                   ],
@@ -62,7 +101,7 @@ class _LocationScreenState extends State<LocationScreen> {
               Padding(
                 padding: EdgeInsets.only(right: 15.0),
                 child: Text(
-                  "It's 🍦 time in San Francisco!",
+                  "$message $cityName!",
                   textAlign: TextAlign.right,
                   style: kMessageTextStyle,
                 ),
